MAINT: renamed src/ fmu/settings/resources to private
