+++ src/fmu/settings/_resources/pydantic_resource_manager.py
